pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wb-transport-hack"
version = "0.1.0"
description = "WB Automated Transportation Dispatch System"
readme = "README.md"
requires-python = ">=3.10"
dependencies =[
    "polars>=1.0.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.4.0",
    "catboost>=1.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.1",
    "flet>=0.21.0",
    "plotly>=5.20.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27,<0.28",
]

# Optional: aliases for easy script execution
[project.scripts]
train = "src.ml_pipeline.train:main"
api = "src.backend.main:start"

[tool.ruff]
line-length = 120
extend-exclude =["notebooks", "data", "models"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["F401"]

[tool.setuptools.packages.find]
exclude = ["data*", "myenv*", "venv*", "env*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
